move css hotreload out of vite.config.js
